diffusion !

+++ talrasha/functional/positional_emb.py
@@ -55,8 +55,8 @@ def tf_sinusoidal_encoding(position: tf.Tensor, d_model, total_step=10000) -> tf
 
     p = position[..., tf.newaxis]
 
-    even_rad = get_angles(p, tf.range(0, d_model, 2)[tf.newaxis, :])
-    odd_rad = get_angles(p, tf.range(1, d_model, 2)[tf.newaxis, :])
+    even_rad = get_angles(p, tf.range(0, d_model, 2, dtype=tf.float32)[tf.newaxis, :])
+    odd_rad = get_angles(p, tf.range(1, d_model, 2, dtype=tf.float32)[tf.newaxis, :])
 
     even_rad = tf.sin(even_rad)
     odd_rad = tf.cos(odd_rad)
